third_party/blink/web_tests/wpt_internal/fenced_frame/csp-fenced-frame-src-allowed.html

<!DOCTYPE html>
<meta http-equiv="Content-Security-Policy" content="fenced-frame-src 'self'">
<title>Test Content-Security-Policy fenced-frame-src</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
const key = KEYS["csp-fenced-frame-src-allowed"];

promise_test(async () => {
  attachFencedFrame("resources/csp-fenced-frame-src-allowed-inner.html");
  const result = await nextValueFromServer(key);
  assert_equals(result, "loaded",
                "The fenced frame is loaded as expected");
}, "csp-fenced-frame-src-allowed");
</script>
</body>
